Cleaning up collective operations

+++ libs/full/segmented_algorithms/tests/unit/partitioned_vector_for_each_double.cpp
@@ -166,7 +166,7 @@ void for_each_tests(std::vector<hpx::id_type>& localities)
 int main()
 {
     std::vector<hpx::id_type> localities = hpx::find_all_localities();
-    for_each_tests<int>(localities);
+    for_each_tests<double>(localities);
     return hpx::util::report_errors();
 }
 #endif
